+++ app/controllers/catalog_controller.rb
@@ -59,6 +59,8 @@ class CatalogController < ApplicationController
 
     @document['published_display'] = "#{@document['published_display'].first}: #{@document['publisher_tesim'].first}" if @document['publisher_tesim'] && @document['published_display']
     @document.delete('publisher_tesim') if @document['published_display']
+
+    @document['oclc'] = HathitrustOverlapRecord.find_by(catalog_id: @document['id'])&.oclc
   end
 
   configure_blacklight do |config|